Change serialization protocol to msgpack (#2)
* Work on mp

* a bit more

* work on mp

* Work om mp

* more mp work

* more mp work

* work on mp

* more mp

* more mp

* work on mp

* more mp

* solved warnings

* fixed first bugs...

* types use fxmap

* Added msgpack print function, fixed warnings, fixed some tests

* fixed many small bugs and tests

* changed proto

* make files

* Improve type instance creation, init thingsdb size

* Fixed unwatch

* Fixed int

+++ inc/ti/fn/fnsplice.h
@@ -127,9 +127,9 @@ static int do__f_splice(ti_query_t * query, cleri_node_t * nd, ex_t * e)
                 task,
                 chain.name,
                 varr,
-                i,
-                c,
-                n))
+                (uint32_t) i,
+                (uint32_t) c,
+                (uint32_t) n))
             goto alloc_err;  /* we do not need to cleanup task, since the task
                                 is added to `query->ev->tasks` */
     }
